Start using ListView instead of manually creating lists
JavaFX has built-in support for displaying lists. This patch changes the code so that it uses ListView, the built-class for displaing a list. This means that whenever you change a small thing, you don't have to recreate the whole UI but only the changes in the list will be changed in the UI. In theory, this should create way better code.