Added Create CardList fully on client side, without server communication
Added a method to create a CardList in CardList class without server side code. The communication with server code has to be added. 

Added AddList to BoardOverviewCtrl Class, this will add a new list to the scene. The button will always stay to the right of the lists.

Added a button to ListView in BoardOverviewCtrl. This button uses addList method in Ctrl.