feat: allow to change order of cards
You can now change the order of cards. Every card has two buttons to move the card up and down. Pressing these buttons moves the card unless it's already at the top/bottom in which case you get an error. When you move a card up or down the list, the pointers of every card to the next card and the pointer of the list to the first card are properly updated. These changes aren't shared with the server though since there's no proper server communication.