Error Fix in code syntax
Old class which has been removed was still in use. Changed it to the new class. BoardListView --> CardListView. 
I also needed to add some extra code because of new attributes in new classes. 

Is the cardList in CardListView and CardListViewCtrl the same object, as in the same memory address? I've coded it like it is.