refactor: move every the hardcoding of board into ServerUtils and prepare for hooking up the api
I moved everything to ServerUtils so in there, it can be replaced with the API call to get the board so we will finally have a basic backend. \n Some noticable detail about the change in commons is that in CardList the @OneToMany annotation is above the getter and not above the field. This is the case cause it would otherwise raise an exception. I'm not sure why, since it works for Board. Querying /board/{id} does give the whole board with the cardlists with the cards so it is functional.