feat: add non-functional popup to edit board and add board info to board overview
This commit adds a popup to edit the board. In the pop-up you will be able to edit the board name and delete the board. You will also be able to see the board id, so you can share it. The board overview has also been changed to display the board id and add a button to get the edit board popup. The fxml files have been slightly changed see the css is more generic. This needs more work in the future.