feat: make list of boards depend on client. Admin can see everything
I added a bunch of features of the backlog. Now, the boards a user sees is independent of other people. You can add a new by board with the add board button. The admin can see every board. You need a password to log in as admin. This password is printed in the log when you run the server. I added automated tests where I could but that's not a lot of places.