fix: fix drag and drop so you can move a card to an empty list
I made it so that moving a card to an empty list works again. The problem was that the event handlers were initialized in CardListView but that wasn't used anymore. They are now moved to CardListViewCtrl. I also made a few other changes/fixes: we know use a special dataformat instead of a string. This is because otherwise you could drag a string like 'c11' from a different application and it would case behaviour here - which should not be the case. I also changed the the transfermode to only use move since that's what we're doing, though I doubt it has any effect on how the application actually works (it's just less confusing). In addition, I fixed a big where moving a card within the same list would put it one card too low if you move it down, with empty spaces in the list as a result (which would cause NullPointerExceptions)